ci: add workflow to record metrics
Summary:
For now, we record and commit, but push with `--dry-run`. This ensures
that the authentication is set up properly without actually making any
changes. Once we want to turn this on, we can just remove the
`--dry-run` flag.

This workflow is set to run on a daily schedule and also on any push to
`master` or to a branch whose name starts with `forcerecord-`, for easy
manual triggering.

Test Plan:
Tested at <https://github.com/sourcecred/metrics/runs/936250395>;
previous attempts, with incorrect authentication, properly failed.